[FLINK-4286] Have Kafka examples that use the Kafka 0.9 connector
