+++ flink-examples/flink-examples-streaming/src/main/java/org/apache/flink/streaming/examples/kafka/WriteIntoKafka09.java
@@ -21,44 +21,56 @@ import org.apache.flink.api.common.restartstrategy.RestartStrategies;
 import org.apache.flink.api.java.utils.ParameterTool;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
-import org.apache.flink.streaming.connectors.kafka.FlinkKafkaConsumer08;
+import org.apache.flink.streaming.api.functions.source.SourceFunction;
+import org.apache.flink.streaming.connectors.kafka.FlinkKafkaProducer09;
 import org.apache.flink.streaming.util.serialization.SimpleStringSchema;
 
-
 /**
- * Read Strings from Kafka and print them to standard out.
- * Note: On a cluster, DataStream.print() will print to the TaskManager's .out file!
+ * Generate a String every 500 ms and write it into a Kafka v0.9 topic.
  *
  * <p>Please pass the following arguments to run the example:
- * 	--topic test --bootstrap.servers localhost:9092 --zookeeper.connect localhost:2181 --group.id myconsumer
+ * 	--topic test --bootstrap.servers localhost:9092
  */
-public class ReadFromKafka {
+public class WriteIntoKafka09 {
 
 	public static void main(String[] args) throws Exception {
-		// parse input arguments
-		final ParameterTool parameterTool = ParameterTool.fromArgs(args);
-
-		if (parameterTool.getNumberOfParameters() < 4) {
-			System.out.println("Missing parameters!\nUsage: Kafka --topic <topic> " +
-					"--bootstrap.servers <kafka brokers> --zookeeper.connect <zk quorum> --group.id <some id>");
+		ParameterTool parameterTool = ParameterTool.fromArgs(args);
+		if (parameterTool.getNumberOfParameters() < 2) {
+			System.out.println("Missing parameters!");
+			System.out.println("Usage: Kafka --topic <topic> --bootstrap.servers <kafka brokers>");
 			return;
 		}
 
 		StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
 		env.getConfig().disableSysoutLogging();
 		env.getConfig().setRestartStrategy(RestartStrategies.fixedDelayRestart(4, 10000));
-		env.enableCheckpointing(5000); // create a checkpoint every 5 seconds
-		env.getConfig().setGlobalJobParameters(parameterTool); // make parameters available in the web interface
 
-		DataStream<String> messageStream = env
-				.addSource(new FlinkKafkaConsumer08<>(
-						parameterTool.getRequired("topic"),
-						new SimpleStringSchema(),
-						parameterTool.getProperties()));
+		// very simple data generator
+		DataStream<String> messageStream = env.addSource(new SourceFunction<String>() {
+			private static final long serialVersionUID = 6369260445318862378L;
+			public boolean running = true;
 
-		// write kafka stream to standard out.
-		messageStream.print();
+			@Override
+			public void run(SourceContext<String> ctx) throws Exception {
+				long i = 0;
+				while (this.running) {
+					ctx.collect("Element - " + i++);
+					Thread.sleep(500);
+				}
+			}
 
-		env.execute("Read from Kafka example");
+			@Override
+			public void cancel() {
+				running = false;
+			}
+		});
+
+		// write data into Kafka
+		messageStream.addSink(new FlinkKafkaProducer09<>(
+			parameterTool.getRequired("topic"),
+			new SimpleStringSchema(),
+			parameterTool.getProperties()));
+
+		env.execute("Write into Kafka v0.9 example");
 	}
 }
